split code tab into separate component and clean up code
